Compatible with Linux-5.4

+++ dax.c
@@ -789,7 +789,7 @@ ssize_t do_nova_inplace_file_write(struct file *filp,
 
 	ret = written;
 	NOVA_STATS_ADD(inplace_write_breaks, step);
-	nova_dbgv("blocks: %lu, %lu\n", inode->i_blocks, sih->i_blocks);
+	nova_dbgv("blocks: %llu, %lu\n", (u64)inode->i_blocks, sih->i_blocks);
 
 	*ppos = pos;
 	if (pos > inode->i_size) {
@@ -902,8 +902,8 @@ static int nova_dax_get_blocks(struct inode *inode, sector_t iblock,
 
 	NOVA_START_TIMING(dax_get_block_t, get_block_time);
 
-	nova_dbgv("%s: pgoff %lu, num %lu, create %d\n",
-				__func__, iblock, max_blocks, create);
+	nova_dbgv("%s: pgoff %llu, num %lu, create %d\n",
+				__func__, (u64)iblock, max_blocks, create);
 
 	epoch_id = nova_get_epoch_id(sb);
 
@@ -920,8 +920,8 @@ again:
 	if (entry) {
 		if (create == 0 || inplace) {
 			nvmm = get_nvmm(sb, sih, entryc, iblock);
-			nova_dbgv("%s: found pgoff %lu, block %lu\n",
-					__func__, iblock, nvmm);
+			nova_dbgv("%s: found pgoff %llu, block %lu\n",
+					__func__, (u64)iblock, nvmm);
 			goto out;
 		}
 	}
